% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosfunc.R
\name{cosfunc}
\alias{cosfunc}
\title{Create a function to compute lagged cosines of velocity vectors}
\usage{
cosfunc(obj = NULL, x0, y0, z0, x1, y1, z1, lag = 0)
}
\arguments{
\item{obj}{A dyadic motion object outputted by \code{\link{make.fd}}.}

\item{x0, y0, z0, x1, y1, z1}{Functional data objects (see \code{\link[fda]{fd}}) representing the three coordinates for persons 0 and 1. These are extracted from \code{obj} if the latter is provided.}

\item{lag}{Lag for person 1 with respect to person 0. E.g., if lag=0.1, person 1 is assumed to lag behind person 0 by 0.1 second.}
}
\value{
A function that inputs a vector of times, and outputs the cosines between the two 3D velocity vectors at each of these times, at the given lag.
}
\description{
Given a dyadic motion object produced by \code{\link{make.fd}}, along with a fixed lag, outputs a function mapping time points to lagged cosine of velocity vectors.
}
\seealso{
\code{\link{multilag}}, which depends on this function. 
Most users would probably use \code{\link{multilag}} rather than calling \code{cosfunc} directly.
}
