% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.fd.R
\name{make.fd}
\alias{make.fd}
\title{Processing of dyadic 3D motion data}
\usage{
make.fd(tvec, p0, p1, nbasis = 300, lambda = NULL)
}
\arguments{
\item{tvec}{Vector of times}

\item{p0, p1}{Matrices of 3D positions for persons 0 and 1 at the times given by \code{tvec}; each should have dimension \code{length(tvec)} by 3}

\item{nbasis}{Dimension of B-spline basis used to represent each person's data along each dimension}

\item{lambda}{Tuning parameter for penalized smoothing with respect to the given B-spline basis.}
}
\value{
A list with elements
\item{x0,y0,z0}{Functional data objects (see \code{\link[fda]{fd}}) representing the three coordinates for person 0} \item{x1,y1,z1}{Same, for person 1}  \item{bsb}{B-spline basis object (see \code{\link[fda]{create.bspline.basis}})} %% ...
}
\description{
Inputs times and corresponding positions in 3D space for two individuals,
and uses the \pkg{fda} package to create functional data objects representing both persons' data with respect to a B-spline basis.
}
\examples{
# see ?syncsham
}
