% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilag.R
\name{multilag}
\alias{multilag}
\title{Compute and plot cosine of velocity vectors at each of a set of lags}
\usage{
multilag(obj, timerange = NULL, lagvec = -2:2/10, npoints = 701,
  power = 1, power.plot = 1, plot = "fields",
  legend.args = list(text = "", side = 3), horizontal = FALSE,
  xlab = "Time", ylab = "Lag", ...)
}
\arguments{
\item{obj}{A dyadic motion object outputted by \code{\link{make.fd}}.}

\item{timerange}{Time range to include in the plot. If \code{NULL}, the entire recorded time range.}

\item{lagvec}{Vector of lag values.}

\item{npoints}{Number of time points, within \code{timerange}, to include in the plots.}

\item{power}{Power to which the computed cosine is raised.}

\item{power.plot}{Power to which the plotted cosine is raised.}

\item{plot}{Either \code{"fields"}, to use \code{\link[fields]{image.plot}} for a color plot of the cosine matrix; or \code{"ggplot"}, to use \code{\link[ggplot2]{ggplot}} to plot the cosines as a function of time for each lag.}

\item{legend.args, horizontal}{Passed to \code{\link[fields]{image.plot}}.}

\item{xlab, ylab}{x- and y-axis labels.}

\item{\dots}{Other arguments to the plotting function.}
}
\value{
A list with components
\item{data.frame}{Data frame used to make the plot} \item{cosmat}{Matrix of cosines} \item{timerange}{The time range supplied}
}
\description{
Inputs a dyadic motion object outputted by \code{\link{make.fd}}, computes cosine of the two individuals' velocity vectors at each of the provided lag values, and makes a plot.
}
\examples{
# Two types of lag plots, as in Figs. 5 and 6 of Reiss, Gvirts et al.
data(sync.fd)
rnge <- c(61,62.2)  

fig5 <- multilag(sync.fd, lagvec=seq(-.4,.4,,31), npoints=801, horizontal=TRUE)
abline(v=rnge[1], col="red")
abline(v=rnge[2], col="red")

fig6 <- multilag(sync.fd, rnge, plot="ggplot")
}
